마이 페이지 리펙토링 완료
